src/tailor-api/src/requirements.txt
Flask>=3.0
Authlib>=1.3
python-dotenv>=1.0
requests>=2.32.5
flask-restful>=0.3.10
pyyaml
psycopg2-binary
